refactor: layout and view
